refactor: pages to Pages
