+++ src/common/constants/developersConst.ts
@@ -1,0 +1,21 @@
+const DEVELOPERS = [
+    {
+        name: 'Николай Козырев',
+        gitHub: 'https://github.com/nikolakozaryan',
+        role: 'developer',
+        contribution: 'В проекте делал...',
+    },
+    {
+        name: 'Яков Григорьев',
+        gitHub: 'https://github.com/YaGrig',
+        role: 'developer',
+        contribution: 'В проекте делал...',
+    },
+    {
+        name: 'Кирилл Найдович',
+        gitHub: 'https://github.com/OnlyR10',
+        role: 'developer',
+        contribution: 'В проекте делал...',
+    },
+];
+export default DEVELOPERS;
