--- src/components/CardTeam/team.ts
@@ -1,9 +1,0 @@
-const team = [
-    {
-        name: 'Николай Козырев',
-        photo: '../../assets/images/photo',
-        role: 'developer',
-        contribution: 'В проекте делал...',
-    },
-];
-export default team;
