+++ src/components/pages2/Statistic/Statistic.tsx
